rename CRABInterface/Utils. Avoid name clash with WMCore/Utils
